+++ azimint/xilinx/Azimint.h
@@ -3,4 +3,4 @@
 constexpr int N = 1024;
 constexpr int npt = 10;
 
-void Azimint(double const *data, double const *radius, double *res);
+void Azimint(float const *data, float const *radius, float *res);
